Move Pagination class to parent package
