+++ src/main/php/de/thekid/dialog/web/Browse.php
@@ -1,6 +1,6 @@
 <?php namespace de\thekid\dialog\web;
 
-use de\thekid\dialog\Repository;
+use de\thekid\dialog\{Repository, Pagination};
 use util\Date;
 use web\frontend\{Handler, Get, Param};
